+++ packages/sit-onyx/src/components/OnyxCheckbox/OnyxCheckbox.vue
@@ -8,6 +8,7 @@ const props = withDefaults(defineProps<OnyxCheckboxProps>(), {
   indeterminate: false,
   disabled: false,
   required: false,
+  truncation: "ellipsis",
   skeleton: false,
 });
 
@@ -44,7 +45,9 @@ const isTouched = ref(false);
         v-model="isChecked"
         :aria-label="props.hideLabel ? props.label : undefined"
         class="onyx-checkbox__input"
-        :class="{ 'onyx-checkbox__input--touched': isTouched }"
+        :class="{
+          'onyx-checkbox__input--touched': isTouched,
+        }"
         type="checkbox"
         :indeterminate="props.indeterminate"
         :disabled="props.disabled"
@@ -53,7 +56,13 @@ const isTouched = ref(false);
       />
     </div>
 
-    <p v-if="!props.hideLabel" class="onyx-checkbox__label">{{ props.label }}</p>
+    <p
+      v-if="!props.hideLabel"
+      class="onyx-checkbox__label"
+      :class="[`onyx-truncation-${props.truncation}`]"
+    >
+      {{ props.label }}
+    </p>
   </label>
 </template>
 
@@ -90,7 +99,8 @@ $input-size: 1rem;
   display: inline-flex;
   align-items: center;
   cursor: pointer;
-  max-width: max-content;
+  width: max-content;
+  max-width: 100%;
 
   &:has(&__label) {
     padding-right: var(--onyx-spacing-2xs);