+++ packages/sit-onyx/src/components/OnyxCheckboxGroup/OnyxCheckboxGroup.vue
@@ -83,6 +83,7 @@ const masterCheckboxState = computed<Partial<OnyxCheckboxProps>>(() => {
   padding: 0;
   border: none;
   max-width: max-content;
+  min-width: unset;
 
   &__label {
     margin-bottom: var(--onyx-spacing-2xs);