+++ packages/sit-onyx/src/components/OnyxRadioButton/OnyxRadioButton.stories.ts
@@ -1,3 +1,4 @@
+import { createTruncationDecorator } from "@/utils/storybook";
 import { defineStorybookActionsAndVModels } from "@sit-onyx/storybook-utils";
 import type { Meta, StoryObj } from "@storybook/vue3";
 import OnyxRadioButton from "./OnyxRadioButton.vue";
@@ -26,6 +27,18 @@ export const Default = {
     name: "radio-name",
     value: "radio-value",
   },
+} satisfies Story;
+
+/**
+ * This example shows a radio truncated radio button.
+ * You can set the "truncation" property to choose between the different truncation types.
+ */
+export const WithTruncation = {
+  args: {
+    ...Default.args,
+    label: "Very long label that will be truncated",
+  },
+  decorators: [createTruncationDecorator("12rem")],
 } satisfies Story;
 
 /**