+++ packages/sit-onyx/CHANGELOG.md
@@ -1,5 +1,17 @@
 # sit-onyx
 
+## 1.0.0-alpha.31
+
+### Minor Changes
+
+- 69f1569: feat: add truncation for OnyxButton, OnyxCheckbox, OnyxRadioButton and OnyxSwitch
+
+## 1.0.0-alpha.30
+
+### Patch Changes
+
+- c785597: fix: add padding-right for checkbox and radio button label
+
 ## 1.0.0-alpha.29
 
 ### Patch Changes